Move DebouncedEvent types into notify-types crate
