+++ notify-debouncer-full/Cargo.toml
@@ -20,6 +20,8 @@ path = "src/lib.rs"
 
 [features]
 default = ["crossbeam","macos_fsevent"]
+serde = ["notify-types/serde"]
+mock_instant = ["dep:mock_instant","notify-types/mock_instant"]
 # can't use dep:crossbeam-channel and feature name crossbeam-channel below rust 1.60
 crossbeam = ["crossbeam-channel","notify/crossbeam-channel"]
 macos_fsevent = ["notify/macos_fsevent"]
@@ -27,14 +29,16 @@ macos_kqueue = ["notify/macos_kqueue"]
 
 [dependencies]
 notify = { version = "6.1.1", path = "../notify", default-features = false }
+notify-types = { version = "1.0.0", path = "../notify-types" }
 crossbeam-channel = { version = "0.5", optional = true }
 file-id = { version = "0.2.1", path = "../file-id" }
 walkdir = "2.2.2"
 log = "0.4.17"
+mock_instant = { version = "0.3.0", optional = true }
 
 [dev-dependencies]
+notify-debouncer-full = { path = ".", features = ["mock_instant"] }
 pretty_assertions = "1.3.0"
-mock_instant = "0.3.0"
 rstest = "0.18"
 serde = { version = "1.0.89", features = ["derive"] }
 deser-hjson = "1.1.1"
