+++ notify-debouncer-mini/Cargo.toml
@@ -20,6 +20,7 @@ path = "src/lib.rs"
 
 [features]
 default = ["crossbeam","macos_fsevent"]
+serde = ["notify-types/serde"]
 # can't use dep:crossbeam-channel and feature name crossbeam-channel below rust 1.60
 crossbeam = ["crossbeam-channel","notify/crossbeam-channel"]
 macos_fsevent = ["notify/macos_fsevent"]
@@ -27,6 +28,6 @@ macos_kqueue = ["notify/macos_kqueue"]
 
 [dependencies]
 notify = { version = "6.1.1", path = "../notify", default-features = false }
+notify-types = { version = "1.0.0", path = "../notify-types" }
 crossbeam-channel = { version = "0.5", optional = true }
-serde = { version = "1.0.89", features = ["derive"], optional = true }
 log = "0.4.17"
