+++ notify-types/Cargo.toml
@@ -16,6 +16,7 @@ edition = "2021"
 
 [dependencies]
 serde = { version = "1.0.89", features = ["derive"], optional = true }
+mock_instant = { version = "0.3.0", optional = true }
 
 [dev-dependencies]
 serde_json = "1.0.39"
