+++ notify-types/src/debouncer_mini.rs
@@ -1,0 +1,34 @@
+use std::path::PathBuf;
+
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
+
+/// A debounced event kind.
+#[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+#[non_exhaustive]
+pub enum DebouncedEventKind {
+    /// No precise events
+    Any,
+    /// Event but debounce timed out (for example continuous writes)
+    AnyContinuous,
+}
+
+/// A debounced event.
+///
+/// Does not emit any specific event type on purpose, only distinguishes between an any event and a continuous any event.
+#[derive(Clone, Debug, Eq, Hash, PartialEq)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
+pub struct DebouncedEvent {
+    /// Event path
+    pub path: PathBuf,
+    /// Event kind
+    pub kind: DebouncedEventKind,
+}
+
+impl DebouncedEvent {
+    #[inline(always)]
+    pub fn new(path: PathBuf, kind: DebouncedEventKind) -> Self {
+        Self { path, kind }
+    }
+}
