+++ notify-types/src/lib.rs
@@ -1,3 +1,5 @@
+pub mod debouncer_full;
+pub mod debouncer_mini;
 pub mod event;
 
 #[cfg(test)]
@@ -25,5 +27,8 @@ mod tests {
         assert_debug_impl!(event::RenameMode);
         assert_debug_impl!(event::Event);
         assert_debug_impl!(event::EventKind);
+        assert_debug_impl!(debouncer_mini::DebouncedEvent);
+        assert_debug_impl!(debouncer_mini::DebouncedEventKind);
+        assert_debug_impl!(debouncer_full::DebouncedEvent);
     }
 }
